services/DocumentProcessing/pyproject.toml
[project]
name = "documentprocessing"
version = "0.1.0"
description = ""
authors = [
    {name = "yithril",email = "yithril@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "transformers (>=4.53.2,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "huggingface-hub (>=0.33.4,<0.34.0)",
    "langchain (>=0.2.0,<0.3.0)",
    "PyPDF2 (>=3.0.0,<4.0.0)",
    "python-docx (>=1.1.0,<2.0.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
